+++ opentelemetry-api/src/opentelemetry/trace/status.py
@@ -48,10 +48,18 @@ class Status:
     ):
         self._status_code = status_code
         self._description = None
+
         if description is not None and not isinstance(description, str):
             logger.warning("Invalid status description type, expected str")
-        else:
-            self._description = description
+            return
+
+        if status_code is not StatusCode.ERROR:
+            logger.warning(
+                "description should only be set when status_code is set to StatusCode.ERROR"
+            )
+            return
+
+        self._description = description
 
     @property
     def status_code(self) -> StatusCode: