+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/open-telemetry/opentelemetry-python/compare/v0.18b0...HEAD)
+- Status now only sets `description` when `status_code` is set to `StatusCode.ERROR`
+  ([#1673](https://github.com/open-telemetry/opentelemetry-python/pull/1673))
 
 ### Added
 - Add `max_attr_value_length` support to Jaeger exporter